refactor(shop): 주소 관련 타입 import 경로 정리
