+++ apps/shop/src/features/order/components/CheckoutView.tsx
@@ -6,7 +6,7 @@ import {
   AddressList,
   DeliverySummary,
 } from '@/features/user/address';
-import type { UserAddress } from '@/features/user/address/types';
+import type { UserAddress } from '@/features/user/address';
 import { CardLayout } from '@/shared/components/layouts';
 import { Button, Modal } from '@findyourkicks/shared';
 import styles from './CheckoutView.module.scss';
